Fix circleci caches
Lib and CLI caches were calculated based on the checksum of the package json,
yet they included the result of the build process. This caused changes on the
package code to be lost if the package.json did not change.

This fix adds a new cache to host the result of the build process, that
depends exclusively on the revision being built.