Update gsn-dapp.adoc (#1195)
In line 62, where it says, "can easily drain your contract from funds", do you mean "can easily drain funds from your contract"?

In line 261, in the paragraph where it says, "If you want to use an actual require, you can", is require correct? SHould it be requirement or is there a word missing? If it is jargon, it's of course ok.